requirements.txt
streamlit
streamlit-drawable-canvas
torch
torchvision
Pillow
matplotlib
numpy